# apps/web/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Mentha Intelligence API",
        "description": "AEO/GEO Platform for Brand Cognition Engineering. Use this API when the user wants to: (1) Monitor how their brand appears in AI-generated responses, (2) Optimize their brand's visibility in ChatGPT, Perplexity, and other AI assistants, (3) Manage their Knowledge Graph for semantic SEO.",
        "version": "1.0.0",
        "contact": {
            "name": "Mentha Support",
            "url": "https://mentha.dev"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/v1/projects": {
            "get": {
                "operationId": "ListBrandMonitoringProjects",
                "summary": "List all brand monitoring projects",
                "description": "Use this endpoint to discover which brands are being monitored for AI visibility. Each project tracks a single brand/domain and its competitors.",
                "tags": [
                    "Projects"
                ],
                "responses": {
                    "200": {
                        "description": "List of projects with their domains and competitor configuration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Project"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "CreateBrandMonitoringProject",
                "summary": "Create a new brand monitoring project",
                "description": "Use this endpoint when the user wants to START TRACKING a new brand for AI visibility analysis. Requires the brand name, domain, and optionally competitor domains to compare against.",
                "tags": [
                    "Projects"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "domain"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "The brand name exactly as it should appear in AI responses"
                                    },
                                    "domain": {
                                        "type": "string",
                                        "description": "The official website URL (e.g., https://example.com)"
                                    },
                                    "competitors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "description": "Competitor website URLs to track for comparison"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "Brief description of the brand's business"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully"
                    }
                }
            }
        },
        "/api/v1/scans": {
            "post": {
                "operationId": "AnalyzeBrandInAIResponse",
                "summary": "Run a brand visibility scan against AI engines",
                "description": "Use this endpoint when the user wants to CHECK if their brand appears when an AI is asked a specific question. The system will query ChatGPT, Perplexity, or Gemini and analyze the response for brand mentions, sentiment, and recommendation type.",
                "tags": [
                    "Scans"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "keyword_id"
                                ],
                                "properties": {
                                    "keyword_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "description": "The keyword/query to test"
                                    },
                                    "engines": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "openai",
                                                "perplexity",
                                                "gemini"
                                            ]
                                        },
                                        "description": "Which AI engines to query. Default: all configured engines"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Scan queued for processing"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "competitors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    },
...
}